refactor: moved constants from utils folder to constatns folder
